app/requirements.txt
flask==2.0.1
werkzeug==2.3.6
mysql-connector-python==8.0.26
psutil==5.8.0
